requirements.txt
requests
flask
pytest
pytest-playwright
tox
